Add files via upload
This version of the meta-analysis has a relevant addition to Section 6 where we looked at whether the decision to combine data from multiple probes via averaging when multiple probes mapped to one gene symbol was biasing our data. We looked at the issue by questioning whether the proportion of gene symbols with multiple probes was similar in the set of all genes vs. the set of genes with statistically significant differential expression. We concluded that the proportions were similar enough that gene symbols that had mapped to more than one probe were not being overrepresented or underrepresented in our final results.